UpcomingDividendsAsync, UpcomingSplitsAsync & UpcomingIposAsync
